examples/otel-value-logging/requirements.txt
shiny
opentelemetry-api
opentelemetry-sdk
